get review is working
