+++ src/server/api/items.js
@@ -1,7 +1,12 @@
 const express = require("express");
 const itemsRouter = express.Router();
 
-const { fetchItems, fetchSingleItem, createItem } = require("../db/items.js");
+const {
+  fetchItems,
+  fetchSingleItem,
+  createItem,
+  fetchReviews,
+} = require("../db/items.js");
 
 itemsRouter.get("/", async (req, res, next) => {
   try {
@@ -37,6 +42,18 @@ itemsRouter.post("/", async (req, res, next) => {
     });
 
     res.status(201).json(newItem);
+  } catch (error) {
+    next(error);
+  }
+});
+
+// this is for getting review for a single item//
+itemsRouter.get("/:id/reviews", async (req, res, next) => {
+  try {
+    const itemId = req.params.id;
+    const getReviews = await fetchReviews(itemId);
+    // res.send(await fetchReviews(req.params.id));
+    res.send({ getReviews });
   } catch (error) {
     next(error);
   }
